+++ crates/gateway-types/src/reply.rs
@@ -2363,9 +2363,8 @@ mod tests {
     use crate::reply::transaction::L1HandlerTransaction;
 
     /// The aim of these tests is to make sure pathfinder is still able to correctly
-    /// deserialize replies from the mainnet sequencer when it still is using some
-    /// previous version of cairo while at the same time the goerli sequencer is
-    /// already using a newer version.
+    /// deserialize replies for some older blocks. The fixtures come from goerli testnet
+    /// which was removed however most or all of the specific cases were also present on mainnet.
     mod backward_compatibility {
         use super::super::StateUpdate;
         use starknet_gateway_test_fixtures::*;