+++ crates/pathfinder/src/state/block_hash.rs
@@ -154,29 +154,11 @@ mod meta {
         }
     }
 
-    const GOERLI_TESTNET_METAINFO: BlockHashMetaInfo = BlockHashMetaInfo {
-        first_0_7_block: BlockNumber::new_or_panic(47028),
-        not_verifiable_range: Some(
-            BlockNumber::new_or_panic(119802)..BlockNumber::new_or_panic(148428),
-        ),
-        fallback_sequencer_address: Some(sequencer_address!(
-            "046a89ae102987331d369645031b49c27738ed096f2789c24449966da4c6de6b"
-        )),
-    };
-
     const MAINNET_METAINFO: BlockHashMetaInfo = BlockHashMetaInfo {
         first_0_7_block: BlockNumber::new_or_panic(833),
         not_verifiable_range: None,
         fallback_sequencer_address: Some(sequencer_address!(
             "021f4b90b0377c82bf330b7b5295820769e72d79d8acd0effa0ebde6e9988bc5"
-        )),
-    };
-
-    const GOERLI_INTEGRATION_METAINFO: BlockHashMetaInfo = BlockHashMetaInfo {
-        first_0_7_block: BlockNumber::new_or_panic(110511),
-        not_verifiable_range: Some(BlockNumber::new_or_panic(0)..BlockNumber::new_or_panic(110511)),
-        fallback_sequencer_address: Some(sequencer_address!(
-            "046a89ae102987331d369645031b49c27738ed096f2789c24449966da4c6de6b"
         )),
     };
 
@@ -201,8 +183,6 @@ mod meta {
     pub fn for_chain(chain: Chain) -> &'static BlockHashMetaInfo {
         match chain {
             Chain::Mainnet => &MAINNET_METAINFO,
-            Chain::GoerliTestnet => &GOERLI_TESTNET_METAINFO,
-            Chain::GoerliIntegration => &GOERLI_INTEGRATION_METAINFO,
             Chain::SepoliaTestnet => &SEPOLIA_TESTNET_METAINFO,
             Chain::SepoliaIntegration => &SEPOLIA_INTEGRATION_METAINFO,
             Chain::Custom => &CUSTOM_METAINFO,
@@ -502,6 +482,7 @@ mod tests {
     use pathfinder_common::felt;
     use pathfinder_common::macro_prelude::*;
     use pathfinder_common::transaction::{EntryPointType, InvokeTransactionV0};
+    use pathfinder_crypto::Felt;
 
     #[test]
     fn test_event_hash() {
@@ -557,28 +538,22 @@ mod tests {
 
     #[test]
     fn test_number_of_events_in_block() {
-        let json = starknet_gateway_test_fixtures::v0_9_0::block::NUMBER_156000;
+        let json = starknet_gateway_test_fixtures::v0_8_0::block::MAINNET_2500;
         let block: Block = serde_json::from_str(json).unwrap();
 
         // this expected value comes from processing the raw JSON and counting the number of events
-        const EXPECTED_NUMBER_OF_EVENTS: usize = 55;
+        const EXPECTED_NUMBER_OF_EVENTS: usize = 26;
         assert_eq!(number_of_events_in_block(&block), EXPECTED_NUMBER_OF_EVENTS);
     }
 
     #[test]
     fn test_block_hash_without_sequencer_address() {
         // This tests with a post-0.7, pre-0.8.0 block where zero is used as the sequencer address.
-        let json = starknet_gateway_test_fixtures::v0_9_0::block::NUMBER_90000;
+        let json = starknet_gateway_test_fixtures::v0_7_0::block::MAINNET_2240;
         let block: Block = serde_json::from_str(json).unwrap();
 
         assert_matches!(
-            verify_block_hash(
-                &block,
-                Chain::GoerliTestnet,
-                ChainId::GOERLI_TESTNET,
-                block.block_hash
-            )
-            .unwrap(),
+            verify_block_hash(&block, Chain::Mainnet, ChainId::MAINNET, block.block_hash).unwrap(),
             VerifyResult::Match(_)
         );
     }
@@ -587,17 +562,11 @@ mod tests {
     fn test_block_hash_with_sequencer_address() {
         // This tests with a post-0.8.2 block where we have correct sequencer address
         // information in the block itself.
-        let json = starknet_gateway_test_fixtures::v0_9_0::block::NUMBER_231579;
+        let json = starknet_gateway_test_fixtures::v0_9_0::block::MAINNET_2800;
         let block: Block = serde_json::from_str(json).unwrap();
 
         assert_matches!(
-            verify_block_hash(
-                &block,
-                Chain::GoerliTestnet,
-                ChainId::GOERLI_TESTNET,
-                block.block_hash
-            )
-            .unwrap(),
+            verify_block_hash(&block, Chain::Mainnet, ChainId::MAINNET, block.block_hash).unwrap(),
             VerifyResult::Match(_)
         );
     }
@@ -607,34 +576,22 @@ mod tests {
         // This tests with a post-0.8.0 pre-0.8.2 block where we don't have the sequencer
         // address in the JSON but the block hash was calculated with the magic value below
         // instead of zero.
-        let json = starknet_gateway_test_fixtures::v0_9_0::block::NUMBER_156000;
+        let json = starknet_gateway_test_fixtures::v0_8_0::block::MAINNET_2500;
         let block: Block = serde_json::from_str(json).unwrap();
 
         assert_matches!(
-            verify_block_hash(
-                &block,
-                Chain::GoerliTestnet,
-                ChainId::GOERLI_TESTNET,
-                block.block_hash,
-            )
-            .unwrap(),
+            verify_block_hash(&block, Chain::Mainnet, ChainId::MAINNET, block.block_hash,).unwrap(),
             VerifyResult::Match(_)
         );
     }
 
     #[test]
     fn test_block_hash_0_11_1() {
-        let json = starknet_gateway_test_fixtures::integration::block::NUMBER_285915;
+        let json = starknet_gateway_test_fixtures::v0_11_1::block::MAINNET_65000;
         let block: Block = serde_json::from_str(json).unwrap();
 
         assert_matches!(
-            verify_block_hash(
-                &block,
-                Chain::GoerliIntegration,
-                ChainId::GOERLI_INTEGRATION,
-                block.block_hash,
-            )
-            .unwrap(),
+            verify_block_hash(&block, Chain::Mainnet, ChainId::MAINNET, block.block_hash,).unwrap(),
             VerifyResult::Match(_)
         );
     }
@@ -643,17 +600,11 @@ mod tests {
     fn test_block_hash_0() {
         // This tests with a pre-0.7 block where the chain ID was hashed into
         // the block hash.
-        let json = starknet_gateway_test_fixtures::v0_9_0::block::GENESIS;
+        let json = starknet_gateway_test_fixtures::pre_0_7_0::block::MAINNET_GENESIS;
         let block: Block = serde_json::from_str(json).unwrap();
 
         assert_matches!(
-            verify_block_hash(
-                &block,
-                Chain::GoerliTestnet,
-                ChainId::GOERLI_TESTNET,
-                block.block_hash
-            )
-            .unwrap(),
+            verify_block_hash(&block, Chain::Mainnet, ChainId::MAINNET, block.block_hash).unwrap(),
             VerifyResult::Match(_)
         );
     }